Changed logic of profiles so now there are two repositories for each profile instead of hashmaps for each profile (change was made according to hometask)
