+++ src/main/java/az/pashabank/ls/services/CustomerService.java
@@ -2,6 +2,7 @@ package az.pashabank.ls.services;
 
 import az.pashabank.ls.entities.CustomerDto;
 import az.pashabank.ls.repository.CustomerRepository;
+import az.pashabank.ls.repository.CustomerRepositoryDev;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
