--- src/main/java/az/pashabank/ls/services/HashMapProvider.java
@@ -1,9 +1,0 @@
-package az.pashabank.ls.services;
-
-import az.pashabank.ls.entities.CustomerDto;
-
-import java.util.HashMap;
-
-public interface HashMapProvider {
-    HashMap<Long, CustomerDto> recieveHashMap();
-}
